refact(start): move boot methods to separate files + cleanup code
